refactor: ProfileDefault record with builder inner class
