+++ common/src/main/java/com/wynntils/features/chat/RevealNicknamesFeature.java
@@ -5,7 +5,7 @@
 package com.wynntils.features.chat;
 
 import com.wynntils.core.consumers.features.Feature;
-import com.wynntils.core.consumers.features.ProfileDefaultBuilder;
+import com.wynntils.core.consumers.features.ProfileDefault;
 import com.wynntils.core.persisted.Persisted;
 import com.wynntils.core.persisted.config.Category;
 import com.wynntils.core.persisted.config.Config;
@@ -30,7 +30,7 @@ public class RevealNicknamesFeature extends Feature {
     private static final String NICKNAME_HOVER_TEXT = "§f%s§7's nickname is §f%s";
 
     public RevealNicknamesFeature() {
-        super(ProfileDefaultBuilder.disabledForAll());
+        super(ProfileDefault.DISABLED);
     }
 
     @Persisted
