+++ common/src/main/java/com/wynntils/features/combat/AutoAttackFeature.java
@@ -6,7 +6,7 @@ package com.wynntils.features.combat;
 
 import com.wynntils.core.components.Models;
 import com.wynntils.core.consumers.features.Feature;
-import com.wynntils.core.consumers.features.ProfileDefaultBuilder;
+import com.wynntils.core.consumers.features.ProfileDefault;
 import com.wynntils.core.persisted.config.Category;
 import com.wynntils.core.persisted.config.ConfigCategory;
 import com.wynntils.mc.event.ArmSwingEvent;
@@ -35,7 +35,7 @@ public class AutoAttackFeature extends Feature {
     private WeaponStatus weaponStatus = WeaponStatus.UNKNOWN;
 
     public AutoAttackFeature() {
-        super(ProfileDefaultBuilder.disabledForAll());
+        super(ProfileDefault.DISABLED);
     }
 
     @SubscribeEvent
