+++ common/src/main/java/com/wynntils/features/debug/ContentBookDumpFeature.java
@@ -10,7 +10,7 @@ import com.google.gson.JsonElement;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Models;
 import com.wynntils.core.consumers.features.Feature;
-import com.wynntils.core.consumers.features.ProfileDefaultBuilder;
+import com.wynntils.core.consumers.features.ProfileDefault;
 import com.wynntils.core.consumers.features.properties.RegisterKeyBind;
 import com.wynntils.core.keybinds.KeyBind;
 import com.wynntils.core.persisted.config.Category;
@@ -76,7 +76,7 @@ public class ContentBookDumpFeature extends Feature {
     private Queue<DumpableActivityInfo> manualTrackingRequired = new LinkedList<>();
 
     public ContentBookDumpFeature() {
-        super(ProfileDefaultBuilder.disabledForAll());
+        super(ProfileDefault.DISABLED);
     }
 
     @SubscribeEvent(priority = EventPriority.LOWEST)
