+++ common/src/main/java/com/wynntils/features/debug/LogItemInfoFeature.java
@@ -9,7 +9,7 @@ import com.mojang.brigadier.tree.LiteralCommandNode;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Models;
 import com.wynntils.core.consumers.features.Feature;
-import com.wynntils.core.consumers.features.ProfileDefaultBuilder;
+import com.wynntils.core.consumers.features.ProfileDefault;
 import com.wynntils.core.consumers.features.properties.RegisterCommand;
 import com.wynntils.core.consumers.features.properties.RegisterKeyBind;
 import com.wynntils.core.keybinds.KeyBind;
@@ -40,7 +40,7 @@ public class LogItemInfoFeature extends Feature {
             Commands.literal("show").executes(this::showCommand).build();
 
     public LogItemInfoFeature() {
-        super(ProfileDefaultBuilder.disabledForAll());
+        super(ProfileDefault.DISABLED);
     }
 
     private void onLogItemInfoPress() {
