+++ common/src/main/java/com/wynntils/features/utilities/AutoSkipCutscenesFeature.java
@@ -5,7 +5,7 @@
 package com.wynntils.features.utilities;
 
 import com.wynntils.core.consumers.features.Feature;
-import com.wynntils.core.consumers.features.ProfileDefaultBuilder;
+import com.wynntils.core.consumers.features.ProfileDefault;
 import com.wynntils.core.persisted.Persisted;
 import com.wynntils.core.persisted.config.Category;
 import com.wynntils.core.persisted.config.Config;
@@ -23,7 +23,7 @@ public class AutoSkipCutscenesFeature extends Feature {
     private final Config<SkipCondition> skipCondition = new Config<>(SkipCondition.ALL);
 
     public AutoSkipCutscenesFeature() {
-        super(ProfileDefaultBuilder.disabledForAll());
+        super(ProfileDefault.DISABLED);
     }
 
     @SubscribeEvent
