+++ common/src/main/java/com/wynntils/features/utilities/TranslationFeature.java
@@ -7,7 +7,7 @@ package com.wynntils.features.utilities;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Services;
 import com.wynntils.core.consumers.features.Feature;
-import com.wynntils.core.consumers.features.ProfileDefaultBuilder;
+import com.wynntils.core.consumers.features.ProfileDefault;
 import com.wynntils.core.persisted.Persisted;
 import com.wynntils.core.persisted.config.Category;
 import com.wynntils.core.persisted.config.Config;
@@ -50,7 +50,7 @@ public class TranslationFeature extends Feature {
             new Config<>(TranslationService.TranslationServices.GOOGLEAPI);
 
     public TranslationFeature() {
-        super(ProfileDefaultBuilder.disabledForAll());
+        super(ProfileDefault.DISABLED);
     }
 
     @SubscribeEvent
